chore: remove github copilot java harness
